# docs/swagger.yaml
definitions:
...
  controllers.CurrentUserResponse:
    properties:
      user:
        $ref: '#/definitions/controllers.PublicUserResponse'
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  controllers.LoginRequest:
    properties:
      captcha_token:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  controllers.ProfileUpdateRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      username:
        type: string
    type: object
  controllers.PublicUserResponse:
    properties:
      age:
        type: integer
      ban_reason:
        type: string
      banned_at:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      last_seen_at:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  controllers.RegisterRequest:
    properties:
      age:
        type: integer
      captcha_token:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  controllers.RegisterResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
  controllers.TokenResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/controllers.PublicUserResponse'
    type: object
  controllers.UploadPhotoResponse:
    properties:
      bucket:
        type: string
      content_type:
        type: string
      key:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login user by email
      parameters:
      - description: Login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login response
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid email/password
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Token generation or database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Login user
      tags:
      - user
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Register success response
          schema:
            $ref: '#/definitions/controllers.RegisterResponse'
        "400":
          description: Invalid request or duplicate data
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register user
      tags:
      - user
...
  /api/v1/me:
    get:
      consumes:
      - application/json
      description: Get the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: Current user response
          schema:
            $ref: '#/definitions/controllers.CurrentUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Update the currently authenticated user's profile. Role, status,
        and ban fields are admin-only and cannot be changed here.
      parameters:
      - description: Profile update request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated current user response
          schema:
            $ref: '#/definitions/controllers.CurrentUserResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - user
...
  /api/v1/uploads/photos:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file to configured S3-compatible object storage
        and return its public URL
      parameters:
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Upload success
          schema:
            $ref: '#/definitions/controllers.UploadPhotoResponse'
        "400":
          description: Invalid file
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "502":
          description: Object storage upload failed
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Object storage is not configured
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload photo image
      tags:
      - uploads
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user by email
      parameters:
      - description: Login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login response
          schema:
            $ref: '#/definitions/controllers.TokenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Invalid email/password
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Token generation or database error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Login user
      tags:
      - user
  /users/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Register success response
          schema:
            $ref: '#/definitions/controllers.RegisterResponse'
        "400":
          description: Invalid request or duplicate data
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "503":
          description: Database is not ready
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register user
      tags:
      - user
swagger: "2.0"